used sharp-quote notation to fix a lambda bug
I thought I was passing a function object to mapchar so that it could
apply the list of numbers to the code in the second
argument. I *think* in reality I was passing a quoted list which
mapchar was nice enough to assume was a function.

Well. At least starting emacs no longer dumps a warning to stdout.