added cperl-mode to interpreter-mode-alist
The above works if you only look at the file suffix - but after
loading, emacs will look at the first line of the file (if
appropriate) and see if there is a hashbang specifying an
interpreter. If that interpreter is in =interpreter-mode-alist=, it
will use the mode specified there.

Since =perl-mode= is the default, Perl scripts starting with the
line =#!/bin/bin/perl= will be associated with that despite
the instructions in =auto-mode-alist=, so we need to add the
mapping =(perl . cperl-mode)= in the =interpreter-mode-alist=.