feat: override user-emacs-directory
The default value assumes that this is `~/.emacs.d/`.

This works in the normal startup sequence, but when testing on github
using the command `emacs --batch --eval '(check-parens)' --script
init.el` it fails since the checked out repo is nested deep in a
directory structure.

By updating it to the actual location, all tests pass.

This is step 1. Next step: replace all instances of sys/emacs-root with
user-emacs-directory.